backend/django/NotesAPI/urls.py
"""
URL configuration for NotesAPI project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/5.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
import pika
import json
from django.contrib import admin
from django.urls import path
from django.db.models.signals import post_delete
from django.contrib.auth.models import User
from django.dispatch import receiver
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]


@receiver(post_delete, sender=User)
def user_deleted_handler(sender, instance, **kwargs):
    connection = pika.BlockingConnection(pika.ConnectionParameters('localhost'))
    channel = connection.channel()

    channel.queue_declare(queue='user.deleted')

    message = json.dumps({'user_id': instance.id})
    channel.basic_publish(exchange='', routing_key='user.deleted', body=message)

    connection.close()
